docker/postgres_start/python_docker/requirements.txt
numpy==1.24.2
psycopg2-binary==2.9.6